Data Pipeline ETL/requirements.txt
sqlparse==0.4.2
psycopg2-binary==2.9.3
SQLAlchemy==1.4.40
pandas==1.4.3
mrjob==0.7.4